Add sending and receiving of stats update
